HTML-755 Add userProperty "clientTimezone" when selecting a location.
Add userProperty "clientTimezone" when selecting a location.
If we only have 1 location, if timezones are enabled and that user don't have the user proprety "clientTimezone" set on the DB, it will force the user to click on that single location, if timezones are not enable then it will skip the location selection screen(if only 1 location are possible for that user).

Update the UiFramework version.